pyproject.toml
[project]
name = "auto-daily"
version = "0.1.0"
description = "macOS window activity logger with Ollama daily report generation"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # macOS-only dependencies (PyObjC)
    "pyobjc-core; sys_platform == 'darwin'",
    "pyobjc-framework-Cocoa; sys_platform == 'darwin'",
    "pyobjc-framework-Quartz; sys_platform == 'darwin'",
    "pyobjc-framework-Vision; sys_platform == 'darwin'",
    "pyobjc-framework-CoreML; sys_platform == 'darwin'",
    "pyobjc-framework-Speech; sys_platform == 'darwin'",
    "pyobjc-framework-AVFoundation; sys_platform == 'darwin'",
    # Cross-platform dependencies
    "httpx",
    "pyyaml",
    "python-dotenv",
    "openai",
    "icalendar",
]

[dependency-groups]
dev = [
    "pytest",
    "pytest-asyncio",
    "pytest-xdist",
    "ty",
]

[project.scripts]
auto-daily = "auto_daily:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/auto_daily"]

[tool.ruff]
target-version = "py312"
line-length = 88
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
]
ignore = [
    "E501",   # line too long (handled by formatter)
]

[tool.ruff.lint.isort]
known-first-party = ["auto_daily"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

# PyObjC modules are dynamically generated and not supported by static type checkers
# These files import macOS-only modules (Quartz, Vision, etc.)
[[tool.ty.overrides]]
include = ["src/auto_daily/ocr.py", "src/auto_daily/system.py", "src/auto_daily/speech.py"]
[tool.ty.overrides.rules]
unresolved-import = "ignore"
unresolved-attribute = "ignore"
